backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pydantic==2.7.1
pydantic-settings==2.2.1
spacy==3.8.2
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.2
scikit-learn==1.4.2
numpy==1.26.4
httpx==0.27.0
aiofiles==23.2.1
python-dateutil==2.9.0
Pillow==10.3.0
email-validator==2.1.1
